The solve function can now recursively make guesses and solve the board from any valid starting position.
An extension method for shuffling a collection has been added.
Extension methods which call a delegate with each result have been removed in favor of those which yield return an IEnumerable.

SudokuCli:
BoardWriter now stores a collection of boards and their solutions, if provided. Stored boards can be registered, unregistered, and loaded, and when unloaded, an empty board will be loaded.
When a board is solved, it's solution is registered.
If a board is being solved, and the BoardWriter has it's solution, it will color cells based on whether they are correct or not.

Added load and unload commands.